loss fix

+++ arc/train_t5.py
@@ -25,7 +25,7 @@ TRAIN_CONFIG = {
     "warmup_steps": 1,
     "eval_freq": 1000,
     "checkpoint_freq": 5000,
-    "dtype": torch.float16,
+    "dtype": torch.float32,
     "max_length": 1024,
     "max_eval_examples": 500
 }
